Move modal files to `autofill` and rename dir to `credentials`
