+++ apps/desktop/src/autofill/modal/credentials/fido2-passkey-exists-icon.ts
