+++ apps/desktop/src/autofill/modal/credentials/fido2-vault.component.html
