+++ apps/desktop/src/autofill/modal/credentials/fido2-vault.component.ts
@@ -23,11 +23,11 @@ import {
 } from "@bitwarden/components";
 import { PasswordRepromptService } from "@bitwarden/vault";
 
+import { DesktopSettingsService } from "../../../platform/services/desktop-settings.service";
 import {
   DesktopFido2UserInterfaceService,
   DesktopFido2UserInterfaceSession,
-} from "../../autofill/services/desktop-fido2-user-interface.service";
-import { DesktopSettingsService } from "../../platform/services/desktop-settings.service";
+} from "../../services/desktop-fido2-user-interface.service";
 
 @Component({
   standalone: true,
